requirements.txt
cffi~=1.16.0
defusedxml~=0.7.1
pip~=24.0
attrs~=23.2.0
cryptography~=42.0.5
Django~=5.0.3
asgiref~=3.8.1
sqlparse~=0.4.4
psycopg2~=2.9.9
colorama~=0.4.6
PyYAML~=6.0.1
numpy~=1.26.4
requests~=2.31.0
python-dotenv~=1.0.1
djangorestframework~=3.15.1
Pygments~=2.17.2